Rename Button directory to CustomButton
