+++ src/components/infocomponents/IntroductionInformation.js
@@ -25,7 +25,7 @@ import { Link } from "react-router-dom";
 
 //DETECT DEVICE
 import { isMobileCustom } from "../../util/customDeviceDetect";
-import { CustomButton } from "../module/Buttons/CustomButton";
+import { CustomButton } from "../module/CustomButtons/CustomButton";
 
 const styles = {
   wrapper: {
