+++ src/components/infocomponents/Welcome.js
@@ -21,7 +21,7 @@ import { Link } from "react-router-dom";
 
 //CHECK DEVICE
 import { isMobileCustom } from "../../util/customDeviceDetect";
-import { CustomButton } from "../module/Buttons/CustomButton";
+import { CustomButton } from "../module/CustomButtons/CustomButton";
 
 const styles = {
   wrapper: {
