+++ src/components/projectComponents/ProjectDialog.js
@@ -38,7 +38,7 @@ import Tabs from "../module/Tabs/Tabs";
 import { ProjectTabData } from "../../data/ProjectTabData";
 
 import "./ProjectDialog.css";
-import { CustomIconButton } from "../module/Buttons/CustomButton";
+import { CustomIconButton } from "../module/CustomButtons/CustomButton";
 
 const Transition = React.forwardRef(function Transition(props, ref) {
   return <Slide direction="up" ref={ref} {...props} />;
